requirements.txt
cachetools==5.3.1
cryptography==41.0.4
dataclasses_json==0.6.1
Flask==3.0.0
frozendict==2.3.8
pidfile==0.1.1
python_pidfile==3.1.1
Requests==2.31.0
ruamel.yaml==0.17.33
waitress==2.1.2
flask_wtf==1.2.1
flask_session==0.5.0